ifaces/qeth: fix gen_conf build, harden vnicc paths, add tests
Move fill_qeth_from_kernel(), apply_qeth() and verify_qeth() out of the
always-compiled schema file ifaces/ethernet.rs into query_apply/ethernet.rs.
The nispor module is gated behind the query_apply feature, so those calls
broke the --no-default-features --features gen_conf build on every arch.
Schema types stay unconditional so YAML parsing works everywhere.

- Run vnicc validation from EthernetInterface::sanitize() so invalid
  desired state is rejected pre-apply, not after the NM checkpoint.
- Treat a missing /sys/bus/ccwgroup/devices as 'not a qeth device'
  instead of an internal error.
- Skip verification of vnicc attributes the kernel does not expose,
  matching apply_vnicc()'s warn-and-skip behaviour; emit a debug log
  when rx_bcast (read-only on OSA) verification is skipped.
- Honor learning-timeout write ordering: the qeth driver only permits
  changing learning_timeout while learning is disabled. Read the current
  learning state, disable it if active, write the timeout, then restore
  learning unless the desired state sets it explicitly.
- Fix the rx_bcast field doc: nmstate attempts the write and warn-and-
  skips on EPERM, it does not refuse to write. Documented as best-effort.
- Place the qeth field after sr_iov in EthernetConfig so hardware-specific
  optional fields group together and show output ordering is unchanged
  for existing sr-iov users.
- Move tests to rust/src/lib/unit_tests/qeth_vnicc.rs, drop the [[test]]
  target, and add tests for the qeth-under-ethernet schema placement.
- Fix all --all-features clippy warnings (let-chains, drop redundant
  qeth.is_some() pre-checks).
